+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Share button added to contact sheet (#1465)
 - Changing the connection details will now clear the contacts and the location message backlog (#1598)
 - Messages now include a random `id` (String) field which can be used by any consumer to correlate and distinguish send/return messages
+- `pubExtendedData` preference renamed to `extendedData` (#1654)
 
 ### Bug fixes